// src/cent-methods.enum.ts
export enum CentMethods {
	Publish = 'publish',
	Broadcast = 'broadcast',
	Subscribe = 'subscribe',
	Unsubscribe = 'unsubscribe',
	Disconnect = 'disconnect',
	Refresh = 'refresh',
	Presence = 'presence',
	PresenceStats = 'presence_stats',
	History = 'history',
	HistoryRemove = 'history_remove',
	Channels = 'channels',
	Info = 'info'
}